rename src/protections/protecitons.ts -> src/protections/ProtectionManager.ts
