+++ src/protections/ProtectionManager.ts
